Role-Based Access Control (RBAC,基於角色的存取控制) for PostControl

+++ backend/controllers/PostController.php
@@ -7,6 +7,7 @@ use common\models\Post;
 use common\models\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
 
@@ -94,6 +95,9 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+    	if (!Yii::$app->user->can('createPost')) {
+    		throw new ForbiddenHttpException('抱歉, 您沒有此操作權限');
+    	}
         $model = new Post();  
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -113,6 +117,9 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+    	if (!Yii::$app->user->can('createPost')) {
+    		throw new ForbiddenHttpException('抱歉, 您沒有此操作權限');
+    	}
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -132,6 +139,9 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+    	if (!Yii::$app->user->can('createPost')) {
+    		throw new ForbiddenHttpException('抱歉, 您沒有此操作權限');
+    	}
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
